requirements.txt
pytest==8.3.3
pytest-cov==6.0.0
allure-python-commons==2.13.5
attrs==24.2.0
colorama==0.4.6
coverage==7.5.3
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
allure-pytest==2.13.5

